refactor : MentorPage 디자인 수정에 따라 구조 수정했습니다
