+++ src/app/mentor/_ui/MentorClient/index.tsx
@@ -4,8 +4,7 @@ import { getUserRoleFromJwt } from "@/utils/jwtUtils";
 
 import MenteePageTabs from "./_ui/MenteePageTabs";
 import MentorFindSection from "./_ui/MentorFindSection";
-import MentorPageTabs from "./_ui/MentorPageTabs";
-import MyMentorSection from "./_ui/MyMentorSection";
+import MentorPage from "./_ui/MentorPage";
 
 import { UserRole } from "@/types/mentor";
 
@@ -17,10 +16,7 @@ const MentorClient = () => {
       {isMentor ? (
         // 멘토페이지
         <>
-          {/* 나의 멘토 - 멘티 탭 및 채팅카드 */}
-          <MentorPageTabs />
-          {/* 나의 멘토 페이지 */}
-          <MyMentorSection />
+          <MentorPage />
         </>
       ) : (
         // 멘티페이지
